fix cast type for pre-031 instances

+++ projects/demo/models/staging/decidim/taxonomies/stg_decidim_taxonomy_filters.sql
@@ -20,12 +20,12 @@
     FROM {{ source('decidim', 'decidim_taxonomy_filters') }}
 {% else %}
     SELECT
-        CAST(NULL AS INTEGER) AS id,
-        CAST(NULL AS INTEGER) AS root_taxonomy_id,
-        CAST(NULL AS INTEGER) AS filter_items_count,
+        CAST(NULL AS BIGINT) AS id,
+        CAST(NULL AS BIGINT) AS root_taxonomy_id,
+        CAST(NULL AS BIGINT) AS filter_items_count,
         CAST(NULL AS TIMESTAMP) AS created_at,
         CAST(NULL AS TIMESTAMP) AS updated_at,
-        CAST(NULL AS INTEGER) AS components_count,
+        CAST(NULL AS BIGINT) AS components_count,
         CAST(NULL AS TEXT) AS name,
         CAST(NULL AS TEXT) AS internal_name,
         CAST(NULL AS JSONB) AS participatory_space_manifests
